[WEB-USER][BE][REFACTOR]: 사용자 모델 및 에러 응답 타입 수정 - passwordHash 필드 필수로 변경, 공통 타입으로 에러 메시지 응답 인터페이스 이동
