+++ apps/web-user/backend/prisma/schema.prisma
@@ -14,9 +14,9 @@ datasource db {
 model User {
   id               String   @id @default(cuid())              // 고유 식별자 (UUID)
   userId           String   @unique @map("user_id")           // 사용자 ID (로그인용)
-  passwordHash     String?  @map("password_hash")             // 암호화된 비밀번호 (선택사항)
+  phone            String   @unique                           // 휴대폰 번호
+  passwordHash     String?  @map("password_hash")             // 암호화된 비밀번호
   name             String?                                    // 사용자 실명 (선택사항)
-  phone            String   @unique                           // 휴대폰 번호 (필수, 중복 불가)
   nickname         String?                                    // 닉네임 (선택사항)
   profileImageUrl  String?  @map("profile_image_url")         // 프로필 이미지 URL (선택사항)
   isVerified       Boolean  @default(false) @map("is_verified") // 휴대폰 인증 여부
