+++ apps/web-user/backend/src/modules/auth/services/user.service.ts
@@ -9,11 +9,8 @@ import {
   CheckPhoneRequestDto,
 } from "@web-user/backend/modules/auth/dto/auth-request.dto";
 import { RegisterDataResponseDto } from "@web-user/backend/modules/auth/dto/auth-data-response.dto";
-import {
-  UserInfo,
-  AvailabilityResponse,
-  CreateUser,
-} from "@web-user/backend/common/types/auth.types";
+import { UserInfo, CreateUser } from "@web-user/backend/common/types/auth.types";
+import { AvailabilityResponse } from "@web-user/backend/common/types/common.types";
 
 /**
  * 사용자 관리 서비스
